Forbidden (403) bypass script completed.
This is a shell script that can be used to bypass the forbidden
page. It uses different payload then can help user/attacker to
bypass the 403 (forbidden) page.

This script uses different types of bypass techniques to bypass 403
and they are 1) HTTP Header Bypass 2) Protocol Based Bypass 3) Port Based Bypass
4) HTTP Method Bypass 5) URL Encoded Bypass